+++ lib/src/main/java/im/status/keycard/applet/ApplicationInfo.java
@@ -1,4 +1,4 @@
-package im.status.keycard.keycard;
+package im.status.keycard.applet;
 
 /**
  * Parses the response from a SELECT command. If the card has not yet received the INIT command the isInitializedCard
