+++ lib/src/main/java/im/status/keycard/applet/KeyPath.java
@@ -1,4 +1,4 @@
-package im.status.keycard.keycard;
+package im.status.keycard.applet;
 
 import java.util.StringTokenizer;
