+++ lib/src/main/java/im/status/keycard/applet/SecureChannelSession.java
@@ -1,4 +1,4 @@
-package im.status.keycard.keycard;
+package im.status.keycard.applet;
 
 import im.status.keycard.io.APDUCommand;
 import im.status.keycard.io.APDUException;
